backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
anthropic==0.40.0
sentence-transformers==3.1.0
faiss-cpu==1.8.0
PyMuPDF==1.24.0
python-multipart==0.0.9
pydantic==2.9.0
numpy>=1.26.0
